refactor(bridge/Block.java) : 블록 Enum 리팩터링
블록 Enum 클래스의 위치를 변경했습니다.

- 블록 Enum 클래스 위치 변경에 따른 User 클래스 반영
