+++ src/main/java/bridge/Block.java
@@ -1,4 +1,4 @@
-package ui;
+package bridge;
 
 
 public enum Block {
